requirements.txt
google-generativeai
onyx-database
numpy
python-dotenv
redis
fastapi
uvicorn
pypdf
python-multipart
